# Bound CSE calls for function inlining 