--- a/src/Inline.cpp
+++ b/src/Inline.cpp
@@ -116,7 +116,7 @@ class Inliner : public IRMutator {
                 body = Let::make(func.name() + "." + func_args[i], args[i], body);
             }
 
-            found = true;
+            found++;
 
             return body;
 
@@ -151,12 +151,12 @@ class Inliner : public IRMutator {
     }
 
     Stmt visit(const Provide *op) override {
-        bool old_found = found;
+        int old_found = found;
 
-        found = false;
+        found = 0;
         Stmt stmt = IRMutator::visit(op);
 
-        if (found) {
+        if (found > 1) {
             stmt = common_subexpression_elimination(stmt);
         }
 
@@ -165,9 +165,9 @@ class Inliner : public IRMutator {
     }
 
 public:
-    bool found;
+    int found = 0;
 
-    Inliner(Function f) : func(f), found(false) {
+    Inliner(Function f) : func(f) {
         internal_assert(f.can_be_inlined()) << "Illegal to inline " << f.name() << "\n";
         validate_schedule_inlined_function(f);
     }
@@ -183,7 +183,7 @@ Stmt inline_function(Stmt s, Function f) {
 Expr inline_function(Expr e, Function f) {
     Inliner i(f);
     e = i.mutate(e);
-    if (i.found) {
+    if (i.found > 1) {
         e = common_subexpression_elimination(e);
     }
     return e;