@abadams 2019-01-24
Passed all relevant tests before I merged master into it, so just merging this to avoid wasting buildbot time.